+++ app/resources_rc.py
@@ -1,0 +1,112 @@
+# Resource object code (Python 3)
+# Created by: object code
+# Created by: The Resource Compiler for Qt version 6.10.2
+# WARNING! All changes made in this file will be lost!
+
+from PySide6 import QtCore
+
+qt_resource_data = b"\
+\x00\x00\x02\xfc\
+<\
+svg xmlns=\x22http:\
+//www.w3.org/200\
+0/svg\x22 viewBox=\x22\
+0 0 24 24\x22 width\
+=\x2224\x22 height=\x2224\
+\x22 fill=\x22none\x22>\x0a \
+ <path d=\x22M9.6 4\
+.2L8.8 6.1C8.33 \
+6.27 7.89 6.47 7\
+.48 6.73L5.53 6.\
+31L4.12 7.72L4.5\
+4 9.67C4.28 10.0\
+8 4.08 10.52 3.9\
+1 10.99L2.02 11.\
+79V13.79L3.91 14\
+.59C4.08 15.06 4\
+.28 15.5 4.54 15\
+.91L4.12 17.86L5\
+.53 19.27L7.48 1\
+8.85C7.89 19.11 \
+8.33 19.31 8.8 1\
+9.48L9.6 21.37H1\
+1.6L12.4 19.48C1\
+2.87 19.31 13.31\
+ 19.11 13.72 18.\
+85L15.67 19.27L1\
+7.08 17.86L16.66\
+ 15.91C16.92 15.\
+5 17.12 15.06 17\
+.29 14.59L19.18 \
+13.79V11.79L17.2\
+9 10.99C17.12 10\
+.52 16.92 10.08 \
+16.66 9.67L17.08\
+ 7.72L15.67 6.31\
+L13.72 6.73C13.3\
+1 6.47 12.87 6.2\
+7 12.4 6.1L11.6 \
+4.2H9.6Z\x22 stroke\
+=\x22#d8d8d8\x22 strok\
+e-width=\x221.5\x22 st\
+roke-linejoin=\x22r\
+ound\x22/>\x0a  <circl\
+e cx=\x2210.6\x22 cy=\x22\
+12.8\x22 r=\x222.6\x22 st\
+roke=\x22#d8d8d8\x22 s\
+troke-width=\x221.5\
+\x22/>\x0a</svg>\x0a\
+\x00\x00\x01\x03\
+<\
+svg xmlns=\x22http:\
+//www.w3.org/200\
+0/svg\x22 viewBox=\x22\
+0 0 24 24\x22 width\
+=\x2224\x22 height=\x2224\
+\x22 fill=\x22none\x22>\x0a \
+ <path d=\x22M3.5 1\
+0.5L12 3.5L20.5 \
+10.5V20H14.75V14\
+.5H9.25V20H3.5V1\
+0.5Z\x22 stroke=\x22#d\
+8d8d8\x22 stroke-wi\
+dth=\x221.8\x22 stroke\
+-linecap=\x22round\x22\
+ stroke-linejoin\
+=\x22round\x22/>\x0a</svg\
+>\x0a\
+"
+
+qt_resource_name = b"\
+\x00\x05\
+\x00o\xa6S\
+\x00i\
+\x00c\x00o\x00n\x00s\
+\x00\x0c\
+\x0b\xdf,\xc7\
+\x00s\
+\x00e\x00t\x00t\x00i\x00n\x00g\x00s\x00.\x00s\x00v\x00g\
+\x00\x08\
+\x068W'\
+\x00h\
+\x00o\x00m\x00e\x00.\x00s\x00v\x00g\
+"
+
+qt_resource_struct = b"\
+\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x00\x00\x00\x02\x00\x00\x00\x02\x00\x00\x00\x02\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x00.\x00\x00\x00\x00\x00\x01\x00\x00\x03\x00\
+\x00\x00\x01\x9c\x81\x8eD6\
+\x00\x00\x00\x10\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
+\x00\x00\x01\x9c\x81\x8ek\xf0\
+"
+
+def qInitResources():
+    QtCore.qRegisterResourceData(0x03, qt_resource_struct, qt_resource_name, qt_resource_data)
+
+def qCleanupResources():
+    QtCore.qUnregisterResourceData(0x03, qt_resource_struct, qt_resource_name, qt_resource_data)
+
+qInitResources()